Refactor Makefile

+++ Makefile
@@ -3,31 +3,33 @@ BIN=.fenv/bin
 virtualenv:
 	if ! [ -d ".fenv" ] ; then virtualenv -p python3 .fenv ; fi
 
-requirements: virtualenv
+keys:
 	chmod 600 ~/.ssh/id_ed25519 ~/.ssh/id_ed25519.pub
-	$(BIN)/pip install -r requirements.txt
-	$(BIN)/ansible-galaxy install -r requirements.yml
 
-set_up_localhost: requirements
+requirements: virtualenv
+	if [ ! -f $(BIN)/ansible-galaxy ]; then $(BIN)/pip install -r requirements.txt; fi
+	if [ ! -d /home/$(USER)/.ansible/collections/ ]; then $(BIN)/ansible-galaxy install -r requirements.yml; fi
+
+set_up_localhost: keys requirements
 	$(BIN)/ansible-playbook playbooks/localhost.yml --ask-become-pass
 
-set_up_localhost.%: requirements
+set_up_localhost.%: keys requirements
 	$(BIN)/ansible-playbook playbooks/localhost.yml --ask-become-pass --tags=$*
 
-bootstrap: virtualenv
+bootstrap: requirements
 	$(BIN)/ansible-playbook playbooks/bootstrap.yml
 
-bootstrap.%: virtualenv
+bootstrap.%: requirements
 	$(BIN)/ansible-playbook playbooks/bootstrap.yml --tags=$*
 
-deploy: virtualenv
+deploy: requirements
 	$(BIN)/ansible-playbook playbooks/cide.yml --ask-become-pass
 
-deploy.%: virtualenv
+deploy.%: requirements
 	$(BIN)/ansible-playbook playbooks/cide.yml --ask-become-pass --tags=$*
 
 # Rebuilds CIDE Electron app and reinstalls it locally.
-cide_app: virtualenv
+cide_app: requirements
 	$(BIN)/ansible-playbook playbooks/cide.yml --ask-become-pass --tags nativefier
 	rm -rf ~/Soft/cide/CIDE-linux-x64
 	mkdir -p ~/Soft/cide/
